remove MAY_BACKLOG flag from requests
cryptodev doesn't implement the backlogging mechanism properly,
since it misses the possibility of sleeping and waiting for
wake-up when the crypto driver below starts working on the
submitted backloggable request.
In case the crypto driver below implements backlogging mechanisms,
this can lead to side-effects such as working on previously-free'ed
data.
This patch removes the MAY_BACKLOG flag from the requests.

Change-Id: Ia3d822b1abfc1a51e2ce3e9682476b2c99d19c5a
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>